Added Qemu vga std video card.  This is not ready for general use.
git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@787 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src/jdos/gui/MainBase.java
@@ -44,7 +44,7 @@ public class MainBase {
         try {Thread.sleep(ms);} catch (Exception e){}
     }
 
-    static public /*Bitu*/int GFX_GetRGB(/*Bit8u*/short red,/*Bit8u*/short green,/*Bit8u*/short blue) {
+    static public /*Bitu*/int GFX_GetRGB(/*Bit8u*/int red,/*Bit8u*/int green,/*Bit8u*/int blue) {
 		return ((blue << 0) | (green << 8) | (red << 16)) | (255 << 24);	
     }
